Convert functions to class for character count to absolute position pdf text references
